New Mixin for Inventory Operations
Besides factorizing the state checks, it should allow wms-inventory
to override all Inventory Operations in one shot
