+++ anyblok_wms_base/core/operation/teleportation.py
@@ -11,7 +11,6 @@ from anyblok.column import Integer
 from anyblok.relationship import Many2One
 
 from anyblok_wms_base.exceptions import (
-    OperationForbiddenState,
     OperationContainerExpected,
     )
 
@@ -23,6 +22,7 @@ Operation = Declarations.Model.Wms.Operation
 @register(Operation)
 class Teleportation(Mixin.WmsSingleInputOperation,
                     Mixin.WmsSingleOutcomeOperation,
+                    Mixin.WmsInventoryOperation,
                     Operation):
     """Inventory Operation to record unexpected change of location for PhysObj.
 
@@ -47,25 +47,13 @@ class Teleportation(Mixin.WmsSingleInputOperation,
     @classmethod
     def check_create_conditions(cls, state, dt_execution,
                                 new_location=None, **kwargs):
-        """Forbid creation with wrong states, check new_location is container.
-
-        :raises: :class:`OperationForbiddenState
-                 <anyblok_wms_base.exceptions.OperationForbiddenState>`
-                 if state is not ``'done'``
-
-        TODO make a common Mixin for all inventory Operations.
+        """Check that new_location is a container.
         """
-        if state != 'done':
-            raise OperationForbiddenState(
-                cls, "Teleportation can exist only in the 'done' state",
-                forbidden=state)
         if new_location is None or not new_location.is_container():
             raise OperationContainerExpected(
                 cls, "new_location field value {offender}",
                 offender=new_location)
-
-        super(Teleportation, cls).check_create_conditions(
-            state, dt_execution, **kwargs)
+        super().check_create_conditions(state, dt_execution, **kwargs)
 
     def after_insert(self):
         """Update :attr:`input` Avatar and create a new one.
